requirements.txt
wget>=3.2
MailTm>=0.0.81
requests>=2.28.1
Faker>=18.11.2
wikipedia>=1.4.0
bs4>=0.0.1
googlesearch-python>=1.2.3
duckduckgo-search>=3.8.3
instaloader>=4.10.2